Add petri net

+++ config.hh
@@ -17,8 +17,8 @@ extern Stat doba;
 
 const int HOUR = 60; // 60m
 
-const double T0 = 0.0;   // Simulation start
-const double T1 = 480; // Simulation end
+const int T0 = 0;   // Simulation start
+const int T1 = 480; // Simulation end
 
 const int SLOW_SHIFT_DURATION = 3 * HOUR; // Slow shift length (default 3h)
 const int FAST_SHIFT_DURATION = 4 * HOUR; // Fast shift length (default 4h)
